Implemented header, sidebar, footer and content elements
