+++ delete.php
@@ -35,6 +35,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Delete element</title>
+    <link rel="stylesheet" href="style.css?<?=date('U')?>">
     <meta name="description" content="Deletes selected element.">
     <meta name="keywords" content="spotify, converter, link">
     <meta name="author" content="We, The People">
@@ -43,14 +44,24 @@
     <meta name="robots" content="index, follow">
 </head>
 <body>
-    <h1>Are you sure you want to delete this element?</h1>
-    <div>
-        <form method="post" action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>">
-            <input type="hidden" name="id" value="<?php echo htmlspecialchars($_GET['id']); ?>">
-            <input type="hidden" name="csrf_token" value="<?php echo $csrf_token; ?>">
-            <input class="back" type="button" value="No" onclick="location.href='index.php'">
-            <input type="submit" name="submit" value="Yes">
-        </form>
+    <?php
+        require 'template/header.html';
+        require 'template/sidebar.html';
+    ?>
+    <div id='content'>
+        <div id="output"></div>
+        <h1>Are you sure you want to delete this element?</h1>
+        <div>
+            <form method="post" action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>">
+                <input type="hidden" name="id" value="<?php echo htmlspecialchars($_GET['id']); ?>">
+                <input type="hidden" name="csrf_token" value="<?php echo $csrf_token; ?>">
+                <input class="back" type="button" value="No" onclick="location.href='index.php'">
+                <input type="submit" name="submit" value="Yes">
+            </form>
+        </div>
+    <?php
+        require 'template/footer.php';
+    ?>
     </div>
 </body>
