+++ edit.php
@@ -60,6 +60,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Edit element</title>
+    <link rel="stylesheet" href="style.css?<?=date('U')?>">
     <meta name="description" content="Edit playlist/link information.">
     <meta name="keywords" content="spotify, converter, link">
     <meta name="author" content="We, The People">
@@ -68,7 +69,13 @@
     <meta name="robots" content="index, follow">
 </head>
 <body>
-    <h1>Edit element</h1>
+    <?php
+        require 'template/header.html';
+        require 'template/sidebar.html';
+    ?>
+    <div id='content'>
+        <div id="output"></div>
+        <h1>Edit element</h1>
     <div>
         <form method="post" action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>">
         <input type="hidden" name="id" value="<?php echo htmlspecialchars($id); ?>">
@@ -83,6 +90,9 @@
             <input class="back" type="button" value="Back" onclick="location.href='index.php'">
             <input type="submit" name="submit" value="Submit">
         </form>
+    <?php
+        require 'template/footer.php';
+    ?>
     </div>
 </body>
