+++ index.php
@@ -7,6 +7,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Spotify Link to Jellyfin maintainer</title>
+    <link rel="stylesheet" href="style.css?<?=date('U')?>">
     <meta name="description" content="Tool which converts spotify link information.">
     <meta name="keywords" content="spotify, converter, link">
     <meta name="author" content="We, The People">
@@ -43,78 +44,34 @@
     </style>
 </head>
 <body>
-    <h1>Spotify Link to Jellyfin maintainer</h1>
-    <p class='refresh'>Next playlist refresh in: <span id="timer">00:00:00</span></p>
-    <table>
-        <thead>
-            <tr>
-                <th>URL</th>
-                <th>Check Frequency</th>
-                <th>Last Check</th>
-                <th>Edit</th>
-                <th>Delete</th>
-            </tr>
-        </thead>
-        <tbody>
-            <?php
-                displayJSONDataToTable();
-            ?>
-        </tbody>
-    </table>
-    <br>
-    <a href="add.php">Add New Entry</a>
-    <a href="download.php">Download Data</a>
-    <a href="lmao.php">HTML form :)</a>
-    <a href="sitemap.html">View Sitemap</a>
-    <a href="#" id="fetchOutputLink">Refresh playlist</a>
-    <br><br>
-    <div id="output"></div>
-    <script>
-         function getOutput() {
-            fetch('fetch_output.php')
-                .then(response => response.text())
-                .then(data => {
-                    document.getElementById('output').innerHTML = data;
-                })
-                .catch(error => console.error('Error:', error));
-        }
-        function updateTimer() {
-            const now = new Date();
-            const currentMinutes = now.getMinutes();
-            const currentSeconds = now.getSeconds();
-            const remainingMinutes = 59 - currentMinutes;
-            const remainingSeconds = 60 - currentSeconds;
-            const formatTime = (value) => {
-                return value < 10 ? `0${value}` : value;
-            };
-
-            if (remainingMinutes === 0 && remainingSeconds === 0) {
-                clearInterval(timerInterval);
-                document.getElementById('timer').textContent = "00:00:00";
-                getOutput();
-
-                setTimeout(updateTimer, 1000); // Restart the timer
-            } else {
-
-                let countdown = `00:00`;
-                // `${formatTime(remainingMinutes)}:${formatTime(remainingSeconds)}`;
-                if (remainingSeconds != 60)
-                    document.getElementById('timer').textContent = `${formatTime(remainingMinutes)}:${formatTime(remainingSeconds)}`;
-                else
-                    document.getElementById('timer').textContent = `${formatTime(remainingMinutes+1)}:00`;
-            }
-        }
-
-        document.getElementById('fetchOutputLink').addEventListener('click', function (event) {
-            event.preventDefault();
-            getOutput();
-        });
-
-        // Initial call to set the timer
-        updateTimer();
-
-        // Update the timer every second
-        const timerInterval = setInterval(updateTimer, 1000);
-    </script>
+    <?php
+        require 'template/header.html';
+        require 'template/sidebar.html';
+    ?>
+    <div id='content'>
+        <div id="output"></div>
+        <table>
+            <thead>
+                <tr>
+                    <th>URL</th>
+                    <th>Check Frequency</th>
+                    <th>Last Check</th>
+                    <th>Edit</th>
+                    <th>Delete</th>
+                </tr>
+            </thead>
+            <tbody>
+                <?php
+                    displayJSONDataToTable();
+                ?>
+            </tbody>
+        </table>
+        <br>
+        
+        <br><br>
+        <?php
+            require 'template/footer.php';
+        ?>
+    </div>
 </body>
 </html>
