--- sitemap.html
@@ -1,22 +1,0 @@
-<html>
-    <head>
-        <title>Sitemap</title>
-        <meta name="description" content="Page which contains all the available pages.">
-        <meta name="keywords" content="spotify, converter, link">
-        <meta name="author" content="We, The People">
-        <meta name="date" content="2023-09-20">
-        <meta name="expiry-date" content="2077-09-20">
-        <meta name="robots" content="index, follow">
-    </head>
-    <body>
-        <h1>Here's all the available pages to be visited on this website:</h1>
-        <ul>
-            <li><a href='index.php'>Main page</a></li>
-            <li><a href='add.php'>Add media</a></li>
-            <li><a href='edit.php'>Edit media</a></li>
-            <li><a href='delete.php'>Delete media</a></li>
-            <li><a href='download.php'>Download configuration</a></li>
-            <li><a href='lmao.php'>All HTML form inputs</a></li>
-        </ul>
-    </body>
-</html>
