+++ style.css
@@ -1,0 +1,47 @@
+body {
+    margin: 0px;
+}
+
+#header {
+    padding: 20px;
+    margin: 0;
+    width: calc(100%-40px);
+    background-color: red;
+}
+
+#header h1, #header p {
+    margin: 0;
+}
+
+#content {
+    margin:0;
+    margin-left:200px;
+    width: calc(100%-200px);
+}
+
+#sidebar {
+    margin: 0;
+    padding: 0;
+    width: 200px;
+    background-color: #f1f1f1;
+    position: fixed;
+    height: 100%;
+    overflow: auto;
+}
+  
+#sidebar a {
+    display: block;
+    color: black;
+    padding: 16px;
+    text-decoration: none;
+}
+
+#sidebar a.active {
+    background-color: #04AA6D;
+    color: white;
+}
+
+#sidebar a:hover:not(.active) {
+    background-color: #333;
+    color: white;
+}
