+++ template/header.html
@@ -1,0 +1,4 @@
+<div id="header">
+    <a class="title" href='index.php'>Spotlister</a>
+    <p class='refresh'>Next playlist refresh in: <span id="timer">00:00:00</span></p>
+</div>
