+++ template/sidebar.html
@@ -1,0 +1,7 @@
+<div id="sidebar">
+    <a href="add.php">Add New Entry</a>
+    <a href="download.php">Download Data</a>
+    <a href="sitemap.php">View Sitemap</a>
+    <a href="#" id="fetchOutputLink">Refresh playlist</a>
+    <a href="lmao.php">HTML form :)</a>
+</div>
